refactor(test): group remaining onboarding coverage
