+++ test/e2e/mock-parity.json
@@ -123,7 +123,7 @@
         "nemoclaw/src/blueprint/runner-identity.test.ts",
         "nemoclaw/src/blueprint/runtime-identity.test.ts",
         "src/lib/onboard/inference-providers/compatible-endpoint-gateway-route.test.ts",
-        "test/blueprint-runtime-identity-lifecycle.test.ts",
+        "test/onboarding/blueprint-runtime-identity-lifecycle.test.ts",
         "test/e2e/support/hosted-inference.test.ts",
         "test/e2e/support/onboard-timeout-contract.test.ts",
         "test/onboarding/onboard-inference-gateway-scope.test.ts"
@@ -197,7 +197,7 @@
     {
       "live": "test/e2e/live/network-policy.test.ts",
       "fast": [
-        "test/validate-blueprint.test.ts",
+        "test/onboarding/validate-blueprint.test.ts",
         "test/e2e/support/e2e-cleanup-resources.test.ts",
         "test/e2e/support/e2e-clients.test.ts"
       ]
