+++ test/helpers/vitest-watch-triggers.ts
@@ -155,11 +155,11 @@ export const vitestWatchTriggerPatterns: VitestWatchTriggerPattern[] = [
   },
   {
     pattern: /(?:^|\/)nemoclaw-blueprint\/policies\/presets\/local-memory\.yaml$/,
-    testsToRun: runTests("test/effective-policy-contracts.test.ts"),
+    testsToRun: runTests("test/onboarding/effective-policy-contracts.test.ts"),
   },
   {
     pattern: /(?:^|\/)nemoclaw-blueprint\/policies\/presets\/claude-code\.yaml$/,
-    testsToRun: runTests("test/effective-policy-contracts.test.ts"),
+    testsToRun: runTests("test/onboarding/effective-policy-contracts.test.ts"),
   },
   {
     pattern: /(?:^|\/)agents\/hermes\/(?:mcp-config-transaction|runtime-config-guard)\.py$/,
