+++ test/onboarding/blueprint-runtime-identity-schema.test.ts
@@ -3,8 +3,8 @@
 
 import { describe, expect, it } from "vitest";
 
-import { isRuntimeIdentityConfig } from "../nemoclaw/src/blueprint/runtime-identity.ts";
-import { compileConfigSchema } from "../scripts/validate-configs.mts";
+import { isRuntimeIdentityConfig } from "../../nemoclaw/src/blueprint/runtime-identity.ts";
+import { compileConfigSchema } from "../../scripts/validate-configs.mts";
 
 const validate = compileConfigSchema("schemas/blueprint.schema.json");
 const baseBlueprint = {
