+++ test/onboarding/openshell-0.0.85-migration-review.test.ts
@@ -5,7 +5,7 @@ import fs from "node:fs";
 import path from "node:path";
 import { describe, expect, it } from "vitest";
 
-const repoRoot = path.resolve(import.meta.dirname, "..");
+const repoRoot = path.resolve(import.meta.dirname, "../..");
 const review = fs.readFileSync(
   path.join(repoRoot, "internal", "security-reviews", "openshell-0.0.85-migration-review.md"),
   "utf8",
